+++ tests/integration/connection_pool_integration_tests.py
@@ -40,16 +40,3 @@ class TestConnectionPooling(unittest.TestCase):
             self.assertListEqual(called, called_again,
                                     'Round robin order wasn\'t preserved')
 
-    def testConnectionSpecified(self):
-        """ When a connection is specified to the constructor, a conenction
-        pool object should be created with that single connection
-        """
-        connection = HttpConnection('http://singleconnectionhost:9200')
-        es = Elastic(connection=connection)
-        with patch('rawes.http_connection.requests.Session.request',
-                MagicMock(return_value=None)) as request:
-            request.return_value = Response()
-            es.get()
-            request.assert_called_with('get',
-                                       'http://singleconnectionhost:9200/',
-                                       timeout=None)